@@ -1,16 +1,15 @@
 name: scroll_bottom_navigation_bar
 description: This package allow hide or show bottom navigation bar while scrolling
-version: 2.1.0+1
+version: 3.0.0
 homepage: https://github.com/EdsonOnildoJR/scroll_bottom_navigation_bar
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: ">=2.6.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  rxdart: ^0.23.1
-  scroll_bars_common: ^1.1.0
+  scroll_bars_common: ^2.0.0
 
 dev_dependencies:
   flutter_test:
