@@ -1,5 +1,6 @@
 library scroll_bottom_navigation_bar;
 
-export 'src/scroll_bottom_navigation_bar_controller.dart';
+export 'package:scroll_bars_common/scroll_bars_common.dart' show Snap;
+
 export 'src/scroll_bottom_navigation_bar.dart';
-export 'src/scroll_body.dart';
+export 'src/scroll_bottom_navigation_bar_controller.dart';
