@@ -1,38 +1,41 @@
-import 'dart:async';
-
 import 'package:flutter/material.dart';
-import 'package:rxdart/rxdart.dart';
 import 'package:scroll_bars_common/scroll_bars_common.dart';
 
-class ScrollBottomNavigationBarController extends ScrollBarsController {
-  ScrollBottomNavigationBarController({
-    ScrollController scrollController,
-    bool snap = true,
-  }) : super(scrollController: scrollController, snap: snap);
+extension ScrollBottomNavigationBarControllerExt on ScrollController {
+  static final _controllers = <int, _ScrollBottomNavigationBarController>{};
+
+  _ScrollBottomNavigationBarController get bottomNavigationBar {
+    if (_controllers.containsKey(this.hashCode)) {
+      return _controllers[this.hashCode];
+    }
+
+    return _controllers[this.hashCode] =
+        _ScrollBottomNavigationBarController(this);
+  }
+}
+
+class _ScrollBottomNavigationBarController extends ScrollBarsController {
+  _ScrollBottomNavigationBarController(ScrollController scrollController)
+      : assert(scrollController != null),
+        super(scrollController);
 
   @override
   double height = kBottomNavigationBarHeight;
 
-  final _page$ = BehaviorSubject<int>.seeded(0);
+  /// Notifier of the active page index
+  final tabNotifier = ValueNotifier<int>(0);
 
-  @Deprecated("pageListener method should been used instead")
-  Stream<int> get stream => pageStream;
+  /// Register a closure to be called when the tab changes
+  void tabListener(Function(int) listener) {
+    tabNotifier.addListener(() => listener(tabNotifier.value));
+  }
 
-  Stream<int> get pageStream => _page$.stream;
+  /// Set a new tab
+  void setTab(int index) => tabNotifier.value = index;
 
-  void changePage(int index) => _page$.add(index);
-
-  StreamSubscription<int> pageListener(
-    void Function(int) onData, {
-    Function onError,
-    void Function() onDone,
-    bool cancelOnError,
-  }) {
-    return _page$.stream.listen(
-      onData,
-      onError: onError,
-      onDone: onDone,
-      cancelOnError: cancelOnError,
-    );
+  @override
+  void dispose() {
+    tabNotifier.dispose();
+    super.dispose();
   }
 }
